Successfully turned sync process into a panel
renamed it to background, since it really just deals with background
data and is ever-present. routed bookmarker through background panel,
works beautifully, and syncing seems to work great on top of it. this is
exciting because there arent 50 background processes anymore, just one.

it can be commanded to load different controllers via port messages.

this is much more efficient than before, and IMO a lot simpler to deal
with.